+++ src/pages/index.js
@@ -2,14 +2,12 @@ import App from './app';
 import ErrorPage from './404';
 import Signup from './signup';
 import Login from './login';
-import Profile from './profile';
-import Upload from './uploadPic';
+import Upload from './upload';
 
 export {
   App,
   ErrorPage,
   Signup,
   Login,
-  Profile,
   Upload
 };
