+++ src/utils/axios.js
@@ -7,8 +7,10 @@ const token = localStorage.getItem('token');
 export default axios.create({
   baseURL: REACT_APP_API_URL,
   headers: {
+    withCredentials: true,
     'x-access-token': `${token}`,
     Authorization: `Bearer ${token}`,
     'x-auth-token': `${token}`,
+    Cookie: `token=${token};`,
   },
 });
