+++ src/utils/upload.js
@@ -1,3 +1,5 @@
+import axios from 'axios';
+
 // eslint-disable-next-line consistent-return
 const uploadImage = async (e, name) => {
   try {
